+++ release_notes.md
@@ -5,10 +5,14 @@
 * Fixed a bug, where handing over a TableView based on a Query restricted
   by another TableView would fail to propagate synchronization status correctly
   (issue #1698)
+* Fixed TableViews that represent backlinks to track the same row, even if that row
+  moves within its table.
 
 ### API breaking changes:
 
-* Lorem ipsum.
+* `TableView::depends_on_deleted_linklist` is now `TableView::depends_on_deleted_object`,
+  and will also return true if the target row of a `TableView` that represents backlinks
+  is deleted.
 
 ### Enhancements: